requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
idna==3.2
lxml==4.6.3
multidict==5.1.0
python-decouple==3.4
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
yarl==1.6.3
